restructure middlewares such that they are categorized by type of function
