+++ lib/handlers/rooms/update_room.go
@@ -1,0 +1,33 @@
+package rooms
+
+import (
+	"github.com/gofiber/fiber/v2"
+	"github.com/lrth06/go-chat/lib/models"
+	"github.com/lrth06/go-chat/lib/utils/config"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func UpdateRoom(c *fiber.Ctx) error {
+	idParam := c.Params("id")
+	id, err := primitive.ObjectIDFromHex(idParam)
+	if err != nil {
+		return c.SendStatus(400)
+	}
+	room := models.Room{}
+	if err := c.BodyParser(&room); err != nil {
+		return c.SendStatus(400)
+	}
+	query := bson.D{{Key: "_id", Value: id}}
+	update := bson.D{{Key: "$set", Value: room}}
+	opts := options.FindOneAndUpdate().SetReturnDocument(options.After)
+	if err := config.ConnDB("Rooms").FindOneAndUpdate(c.Context(), query, update, opts).Decode(&room); err != nil {
+		return c.SendStatus(400)
+	}
+
+	return c.JSON(fiber.Map{
+		"msg":  "Room updated successfully!",
+		"room": room,
+	})
+}
